Software funciona casi en su totalidad con un problema aún por resolver (devoluciones)

+++ src/main/java/com/pharmacore/pharmacore/repository/CategoriasRepository.java
@@ -3,7 +3,7 @@ package com.pharmacore.pharmacore.repository;
 import com.pharmacore.pharmacore.model.Categorias;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriasRepository extends JpaRepository<Categorias, Integer>
+public interface CategoriasRepository extends JpaRepository<Categorias, Long>
 {
 
 }
